# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Instagram-clone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  username: string;
  fullName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  postsCount: number;
  followersCount: number;
  followingCount: number;
  isFollowing: boolean;
  isMe: boolean;
  createdAt: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  pronouns?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  interests: string[];
  profileCompleted: boolean;
  isBlocked: boolean;
}