# lib/api-zod/src/generated/types/userReportInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Instagram-clone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserReportInput {
  reason: string;
  details?: string;
}